Added implementation file
- Removed riostream
